requirements.txt
google-cloud-bigquery==0.31.0
google-auth-oauthlib==0.2.0
pytz==2018.3
django==2.0.6
